classを移動
